Refactored API routes into controllers
