+++ src/server/api/controllers/jobs.js
@@ -1,0 +1,13 @@
+export default class JobsController {
+  constructor (workable) {
+    this.workable = workable;
+  }
+
+  getJobs = (req, res) => {
+    this.workable.getJobs()
+      .then(res.send.bind(res))
+      .catch((err) => {
+        res.status(err.status).send(err.message);
+      });
+  };
+}
