+++ src/server/api/index.js
@@ -1,5 +1,6 @@
 import express from 'express';
-import Routes from './routes';
+import JobsController from './controllers/jobs';
+import EventsController from './controllers/events';
 import bodyParser from 'body-parser';
 import authSetup from '../authSetup';
 
@@ -10,16 +11,17 @@ const ensureAuthenticated = (req, res, next) => {
 
 export default function API (workable) {
   const self = authSetup(express());
-  const routes = new Routes(workable);
+  const jobsController = new JobsController(workable);
+  const eventsController = new EventsController();
 
   self.use( bodyParser.json() );       // to support JSON-encoded bodies
   self.use( bodyParser.urlencoded({     // to support URL-encoded bodies
     extended: true
   }));
 
-  self.get('/jobs', routes.getJobs);
-  self.get('/events', routes.getEvents);
-  self.post('/new-event', ensureAuthenticated, routes.postNewEvent);
+  self.get('/jobs', jobsController.getJobs);
+  self.get('/events', eventsController.getEvents);
+  self.post('/new-event', ensureAuthenticated, eventsController.postNewEvent);
   self.get('/check-login', ensureAuthenticated, (req, res) => res.send('OK'));
 
   return self;
